Using part and part of in flutter
The "part" statement is used to run code within one file into another file. This can make the file of a large file more manageable, allowing you to split the code into smaller chunks.
On the other hand, the phrase "Part of" is used to indicate which class or module a file belongs to. This allows you to collect all the code of a particular class or module in a single file, making the code more organized and easily manageable.